+++ app/resources/api/components/BundlesSidebarContent.tsx
@@ -1,28 +1,7 @@
 import Accordion from '@/components/Accordion/Accordion';
 import AccordionGroup from '@/components/Accordion/AccordionGroup';
-import { Bundle, BundleItem } from '@/types/api';
-import styles from '@/styles/accordion.module.scss'
-import Link from 'next/link';
-import slugify from 'slugify';
-
-const renderAccordionContent = (
-  bundles: Array<BundleItem>, baseHref: string
-) => {
-  return (
-    <ul className={styles.accordionMenu}>
-      {bundles?.map((item, index) => {
-        const slug = slugify(item.name)
-        return <li key={slug + '_' + index}>
-          <Link
-            href={baseHref + slug}
-          >
-            {item.name}
-          </Link>
-        </li>
-      })}
-    </ul>
-  )
-}
+import { Bundle } from '@/types/api';
+import SidebarAccordionContent from './SideBarAccordionContent';
 
 export default function BundlesSidebarContent(
   {
@@ -37,13 +16,13 @@ export default function BundlesSidebarContent(
     return <div style={{marginTop: '2em'}}>
       <h3 style={{ fontSize: '1.125rem'}}>Select bundle</h3>
       <AccordionGroup>
-        {elements?.map((element: Bundle) => (
-          <Accordion
+        {elements?.map((element: Bundle) => {
+          return <Accordion
             key={element.name}
             title={ element.name }
-            content={renderAccordionContent(element.bundles, baseHref)}
-          />
-        ))}
+            content={<SidebarAccordionContent elements={element.bundles} baseHref={baseHref} />}
+          />;
+        })}
       </AccordionGroup>;
     </div>
   }
