+++ app/resources/api/components/EventsAndRequestsSidebarContent.tsx
@@ -1,28 +1,7 @@
-import slugify from 'slugify';
 import { OskariEventOrRequest } from '@/types/api';
 import AccordionGroup from '@/components/Accordion/AccordionGroup';
 import Accordion from '@/components/Accordion/Accordion';
-import Link from 'next/link';
-import styles from '@/styles/accordion.module.scss'
-
-const renderAccordionContent = (
-  elements: Array<OskariEventOrRequest>, baseHref: string
-) => {
-  return (
-    <ul className={styles.accordionMenu}>
-      {elements?.map((item, index) => {
-        const slug = slugify(item.name)
-        return <li key={slug + '_' + index}>
-          <Link
-            href={baseHref + slug}
-          >
-            {item.name}
-          </Link>
-        </li>
-      })}
-    </ul>
-  )
-}
+import SidebarAccordionContent from './SideBarAccordionContent';
 
 export default function EventsAndRequestsSidebarContent(
 {
@@ -42,10 +21,11 @@ export default function EventsAndRequestsSidebarContent(
     <h3 style={{ fontSize: '1.125rem'}}>{title}</h3>
     <AccordionGroup>
     { namespaces.map((namespace: string) => {
+        const filtered = elements?.filter((element) => element.ns === namespace);
         return <Accordion
           key={namespace}
           title={ namespace }
-          content={renderAccordionContent(elements?.filter((element) => element.ns === namespace), baseHref)}
+          content={<SidebarAccordionContent elements={filtered} baseHref={baseHref}/>}
         />;
       })
     }
