+++ types/api.ts
@@ -16,3 +16,7 @@ export type OskariEventOrRequest = {
   bundle: string,
   ns: string
 }
+
+export type SidebarAccordionItem = {
+  name: string
+}
